+++ crates/torii/graphql/src/tests/mod.rs
@@ -2,3 +2,4 @@ mod common;
 mod components_test;
 mod entities_test;
 mod events_test;
+mod subscription_test;
